pyproject.toml
[project]
name = "shellbot"
version = "2.0.0"
dependencies = [
    "msgpack>=1.1.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "setuptools>=80.9.0",
]

[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
shellbot = "shellbot.__main__:main"



